update all (-app, -controller, +client interface)
