+++ pom.xml
@@ -97,18 +97,3 @@
         </plugins>
     </build>
 </project>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
