--- src/main/java/dpa/weather/weatherstarter/WeatherStarterApplication.java
@@ -1,13 +1,0 @@
-package dpa.weather.weatherstarter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WeatherStarterApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WeatherStarterApplication.class, args);
-    }
-
-}
