--- src/main/java/dpa/weather/weatherstarter/controller/WeatherController.java
@@ -1,24 +1,0 @@
-package dpa.weather.weatherstarter.controller;
-
-import dpa.weather.weatherstarter.responce.WeatherResponse;
-import dpa.weather.weatherstarter.service.WeatherService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/weather")
-@RequiredArgsConstructor
-public class WeatherController {
-
-    private final WeatherService weatherService;
-
-    @GetMapping("/{city}")
-    public ResponseEntity<WeatherResponse> getWeather(@PathVariable String city) {
-        WeatherResponse response = weatherService.getWeatherForCity(city);
-        return ResponseEntity.ok(response);
-    }
-}
