+++ src/main/java/dpa/weather/weatherstarter/exception/WeatherServiceException.java
