+++ src/main/java/dpa/weather/weatherstarter/properties/WeatherProperties.java
@@ -13,4 +13,3 @@ public class WeatherProperties {
     private String apiUrl;
     private String apiKey;
 }
-
