+++ src/main/java/dpa/weather/weatherstarter/service/WeatherService.java
@@ -1,6 +1,6 @@
 package dpa.weather.weatherstarter.service;
 
-import dpa.weather.weatherstarter.responce.WeatherResponse;
+import dpa.weather.weatherstarter.response.WeatherResponse;
 
 public interface WeatherService {
     WeatherResponse getWeatherForCity(String city);
