+++ src/main/java/dpa/weather/weatherstarter/service/implementation/WeatherServiceImpl.java
@@ -1,7 +1,7 @@
 package dpa.weather.weatherstarter.service.implementation;
 
-import dpa.weather.weatherstarter.client.WeatherClient;
-import dpa.weather.weatherstarter.responce.WeatherResponse;
+import dpa.weather.weatherstarter.client.implementation.WeatherClientImpl;
+import dpa.weather.weatherstarter.response.WeatherResponse;
 import dpa.weather.weatherstarter.service.WeatherService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class WeatherServiceImpl implements WeatherService {
 
-    private final WeatherClient weatherClient;
+    private final WeatherClientImpl weatherClientImpl;
 
     @Override
     public WeatherResponse getWeatherForCity(String city) {
-        return weatherClient.getWeatherByCity(city);
+        return weatherClientImpl.getWeatherByCity(city);
     }
 }
