+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
 spring.application.name=weather-starter
 weather.enabled=true
 weather.api-key=5f8f1e322944e656897cbd9549859b6b
-weather.api-url=https://api.openweathermap.org/data/2.5/weather
+weather.api-url=https://api.openweathermap.org/data/2.5/weather/
