--- src/test/java/dpa/weather/weatherstarter/WeatherStarterApplicationTests.java
@@ -1,13 +1,0 @@
-package dpa.weather.weatherstarter;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WeatherStarterApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
